chat_text_analysis_module/requirements.txt
streamlit==1.40.2
streamlit-authenticator==0.4.1
firebase-admin==6.6.0
PyYAML==6.0.2
langchain-openai==0.2.10
openai==1.55.3